Fix Issue #14 - Connect Normal Map Correctly
When generating a material from baked textures, the normal map image is assigned to the Alpha BSDF node instead of the expected Normal Node. 

I'm guessing the BSDF node was updated by blender to include one more Node input, so that referring to the node input by index (19) broke. 

There is probably some more fundamental fix that can refer to node inputs by name, but that's not included here.