Workaround for Diffuse Baking
The Principled BSDF shader doesn't bake the Diffuse (Base Color) properly when the metallic is not zero. This workaround is for using a Custom Pass with the property set to Base Color. Did some testing and is working for me when baking out PBR work flow (Base Color, Metallic, Roughness, Normal).